v1.0.0

+++ orion-ops-module-infra/orion-ops-module-infra-service/src/main/java/com/orion/ops/module/infra/service/impl/DataExtraServiceImpl.java
@@ -2,9 +2,14 @@ package com.orion.ops.module.infra.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.orion.lang.function.Functions;
+import com.orion.lang.utils.collect.Lists;
 import com.orion.lang.utils.collect.Maps;
 import com.orion.ops.framework.common.constant.Const;
+import com.orion.ops.framework.mybatis.core.query.ThenLambdaWrapper;
+import com.orion.ops.framework.redis.core.utils.RedisMaps;
+import com.orion.ops.framework.redis.core.utils.barrier.CacheBarriers;
 import com.orion.ops.module.infra.dao.DataExtraDAO;
+import com.orion.ops.module.infra.define.cache.DataExtraCacheKeyDefine;
 import com.orion.ops.module.infra.entity.domain.DataExtraDO;
 import com.orion.ops.module.infra.entity.request.data.DataExtraQueryRequest;
 import com.orion.ops.module.infra.entity.request.data.DataExtraSetRequest;
@@ -15,6 +20,7 @@ import org.springframework.stereotype.Service;
 import javax.annotation.Resource;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 /**
@@ -62,15 +68,29 @@ public class DataExtraServiceImpl implements DataExtraService {
         insert.setItem(request.getItem());
         insert.setValue(request.getValue());
         dataExtraDAO.insert(insert);
+        // 删除缓存
+        RedisMaps.delete(DataExtraCacheKeyDefine.DATA_EXTRA.format(request.getUserId(), request.getType(), request.getItem()));
         return insert.getId();
     }
 
     @Override
     public Integer updateExtraValue(Long id, String value) {
+        // 查询数据
+        DataExtraDO data = this.getCacheSelectWrapper()
+                .eq(DataExtraDO::getId, id)
+                .then()
+                .get();
+        if (data == null) {
+            return Const.N_0;
+        }
         DataExtraDO update = new DataExtraDO();
         update.setId(id);
         update.setValue(value);
-        return dataExtraDAO.updateById(update);
+        // 更新
+        int effect = dataExtraDAO.updateById(update);
+        // 删除缓存
+        RedisMaps.delete(DataExtraCacheKeyDefine.DATA_EXTRA.format(data.getUserId(), data.getType(), data.getItem()));
+        return effect;
     }
 
     @Override
@@ -78,8 +98,16 @@ public class DataExtraServiceImpl implements DataExtraService {
         if (Maps.isEmpty(map)) {
             return;
         }
+        // 查询数据
+        List<DataExtraDO> list = this.getCacheSelectWrapper()
+                .in(DataExtraDO::getId, map.keySet())
+                .then()
+                .list();
+        if (list.isEmpty()) {
+            return;
+        }
         // 批量更新
-        List<DataExtraDO> list = map.entrySet()
+        List<DataExtraDO> update = map.entrySet()
                 .stream()
                 .map(s -> {
                     DataExtraDO extra = new DataExtraDO();
@@ -87,7 +115,9 @@ public class DataExtraServiceImpl implements DataExtraService {
                     extra.setValue(s.getValue());
                     return extra;
                 }).collect(Collectors.toList());
-        dataExtraDAO.updateBatch(list);
+        dataExtraDAO.updateBatch(update);
+        // 删除缓存
+        this.deleteCache(list);
     }
 
     @Override
@@ -111,6 +141,36 @@ public class DataExtraServiceImpl implements DataExtraService {
     }
 
     @Override
+    public String getExtraItemValueByCache(Long userId, String type, String item, Long relId) {
+        return this.getExtraItemValuesByCache(userId, type, item).get(relId);
+    }
+
+    @Override
+    public Map<Long, String> getExtraItemValuesByCache(Long userId, String type, String item) {
+        // 查询缓存
+        String key = DataExtraCacheKeyDefine.DATA_EXTRA.format(userId, type, item);
+        Map<String, String> entities = RedisMaps.entities(key);
+        if (Maps.isEmpty(entities)) {
+            // 查询数据库
+            DataExtraQueryRequest request = DataExtraQueryRequest.builder()
+                    .userId(userId)
+                    .type(type)
+                    .item(item)
+                    .build();
+            Map<Long, String> extras = this.getExtraItemValues(request);
+            entities = Maps.map(extras, String::valueOf, String::valueOf);
+            // 设置屏障 防止穿透
+            CacheBarriers.MAP.check(entities);
+            // 设置缓存
+            RedisMaps.putAll(key, DataExtraCacheKeyDefine.DATA_EXTRA, entities);
+        }
+        // 删除屏障
+        CacheBarriers.MAP.remove(entities);
+        // 转换
+        return Maps.map(entities, Long::valueOf, Function.identity());
+    }
+
+    @Override
     public DataExtraDO getExtraItem(DataExtraQueryRequest request) {
         return dataExtraDAO.of()
                 .wrapper(this.buildWrapper(request))
@@ -126,12 +186,62 @@ public class DataExtraServiceImpl implements DataExtraService {
 
     @Override
     public Integer deleteByUserId(Long userId) {
-        return dataExtraDAO.deleteByUserId(userId);
+        List<DataExtraDO> list = this.getCacheSelectWrapper()
+                .eq(DataExtraDO::getUserId, userId)
+                .then()
+                .list();
+        if (list.isEmpty()) {
+            return Const.N_0;
+        }
+        // 删除数据
+        int effect = dataExtraDAO.deleteByUserId(userId);
+        // 删除缓存
+        this.deleteCache(list);
+        return effect;
     }
 
     @Override
     public Integer deleteByRelId(String type, Long relId) {
-        return dataExtraDAO.deleteByRelId(type, relId);
+        List<DataExtraDO> list = this.getCacheSelectWrapper()
+                .eq(DataExtraDO::getType, type)
+                .eq(DataExtraDO::getRelId, relId)
+                .then()
+                .list();
+        if (list.isEmpty()) {
+            return Const.N_0;
+        }
+        // 删除数据
+        int effect = dataExtraDAO.deleteByRelId(type, relId);
+        // 删除缓存
+        this.deleteCache(list);
+        return effect;
+    }
+
+    /**
+     * 获取查询缓存参数 wrapper 不查询 longtext 加速查询
+     *
+     * @return wrapper
+     */
+    private ThenLambdaWrapper<DataExtraDO> getCacheSelectWrapper() {
+        return dataExtraDAO.of()
+                .createWrapper()
+                .select(DataExtraDO::getId, DataExtraDO::getUserId, DataExtraDO::getType, DataExtraDO::getItem);
+    }
+
+    /**
+     * 删除缓存
+     *
+     * @param list list
+     */
+    private void deleteCache(List<DataExtraDO> list) {
+        if (Lists.isEmpty(list)) {
+            return;
+        }
+        List<String> keys = list.stream()
+                .map(s -> DataExtraCacheKeyDefine.DATA_EXTRA.format(s.getUserId(), s.getType(), s.getItem()))
+                .distinct()
+                .collect(Collectors.toList());
+        RedisMaps.delete(keys);
     }
 
     /**
